Moved Fragment Module to correct package
