+++ app/src/main/java/com/github/talkbacktutorial/activities/modules/OpenRecentAppsModuleActivity.kt
@@ -5,9 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.commit
 import com.github.talkbacktutorial.R
-import com.github.talkbacktutorial.activities.lesson0.Lesson0Part1Fragment
 import com.github.talkbacktutorial.databinding.ActivityLesson0Binding
-import com.github.talkbacktutorial.lessons.OpenRecentAppsModuleFragment
 
 class OpenRecentAppsModuleActivity : AppCompatActivity() {
